+++ tests/integration_tests/text_attr.rs
@@ -109,12 +109,18 @@ fn test_text_multiline() {
     let input = r#"
 <rect xy="0" wh="10" text-lsp="1" text="multi\n\n\nline"/>
 "#;
-    let expected = r#"
+    // Slightly hacky, we insert zero-width spaces in empty lines
+    let expected = format!(
+        "{}\u{200b}{}\u{200b}{}",
+        r#"
 <rect x="0" y="0" width="10" height="10"/>
 <text x="5" y="5" class="d-tbox">
-<tspan x="5" dy="-1.5em">multi</tspan><tspan x="5" dy="1em"></tspan><tspan x="5" dy="1em"></tspan><tspan x="5" dy="1em">line</tspan>
+<tspan x="5" dy="-1.5em">multi</tspan><tspan x="5" dy="1em">"#,
+        r#"</tspan><tspan x="5" dy="1em">"#,
+        r#"</tspan><tspan x="5" dy="1em">line</tspan>
 </text>
-"#;
+"#
+    );
     assert_eq!(
         transform_str_default(input).unwrap().trim(),
         expected.trim()